+++ modules/services/repositories/src/test/java/au/com/shiftyjelly/pocketcasts/repositories/playback/Media3SessionCallbackTest.kt
@@ -356,6 +356,7 @@ class Media3SessionCallbackTest {
         verify(playbackManager).pauseSuspend(
             transientLoss = any(),
             sourceView = any(),
+            dueToInterruption = any(),
         )
     }